+++ source/refined-github.ts
@@ -220,4 +220,4 @@ import './features/submission-via-ctrl-enter-everywhere';
 import './features/linkify-user-labels';
 import './features/repo-avatars';
 import './features/jump-to-conversation-close-event';
-import './features/linkify-notification-page-numbers';
+import './features/last-notification-page-button';
